fix: #75 - 리뷰 내용 적용
