+++ src/lib/auth/checkResendRateLimit.ts
@@ -1,37 +1,12 @@
-export type RateLimitResult = {
-  allowed: boolean;
-};
+import {
+  checkLimit,
+  RateLimitResult,
+  WindowEntry,
+} from "@/lib/auth/rateLimitUtils";
 
 const EMAIL_LIMIT = 5;
-const WINDOW_MS = 60 * 1000;
-
-type WindowEntry = {
-  count: number;
-  windowStart: number;
-};
 
 const emailStore = new Map<string, WindowEntry>();
-
-function checkLimit(
-  store: Map<string, WindowEntry>,
-  key: string,
-  limit: number,
-): boolean {
-  const now = Date.now();
-  const entry = store.get(key);
-
-  if (!entry || now - entry.windowStart >= WINDOW_MS) {
-    store.set(key, { count: 1, windowStart: now });
-    return true;
-  }
-
-  if (entry.count >= limit) {
-    return false;
-  }
-
-  entry.count += 1;
-  return true;
-}
 
 export function resetResendRateLimitStore(): void {
   emailStore.clear();
