+++ src/lib/auth/getUserByEmail.ts
@@ -4,13 +4,24 @@ export async function getUserByEmail(
   email: string,
 ): Promise<{ email: string; email_confirmed_at: string | null } | null> {
   const supabase = createAdminClient();
-  const { data } = await supabase.auth.admin.listUsers();
-  const user = data.users.find((u) => u.email === email);
+  let page = 1;
+  const perPage = 1000;
 
-  if (!user) return null;
+  while (true) {
+    const { data } = await supabase.auth.admin.listUsers({ page, perPage });
+    const user = data.users.find((u) => u.email === email);
 
-  return {
-    email: user.email ?? email,
-    email_confirmed_at: user.email_confirmed_at ?? null,
-  };
+    if (user) {
+      return {
+        email: user.email ?? email,
+        email_confirmed_at: user.email_confirmed_at ?? null,
+      };
+    }
+
+    if (data.users.length < perPage) {
+      return null;
+    }
+
+    page += 1;
+  }
 }
