+++ src/lib/auth/resendVerificationEmail.ts
@@ -1,3 +1,4 @@
+// TODO: Supabase admin API를 사용해 인증 이메일 재전송 구현 필요
 export async function resendVerificationEmail(_email: string): Promise<void> {
   throw new Error("Not implemented");
 }
