+++ src/lib/constants/storageBuckets.ts
@@ -1,0 +1,3 @@
+export const STORAGE_BUCKETS = {
+  AVATARS: "avatars",
+} as const;
